Refactor service worker
