+++ frontend/src/features/map/layers/BaseLayer.tsx
@@ -5,6 +5,7 @@ import XYZ from 'ol/source/XYZ'
 import React, { useCallback, useEffect, useMemo, useRef } from 'react'
 
 import { LayerProperties } from '../../../domain/entities/layers/constants'
+import { useIsInNavigationMode } from '../../../hooks/authorization/useIsInNavigationMode'
 import { useMainAppSelector } from '../../../hooks/useMainAppSelector'
 
 import type { ImageTile } from 'ol'
@@ -14,6 +15,7 @@ export type BaseLayerProps = {
   map?: any
 }
 function UnmemoizedBaseLayer({ map }: BaseLayerProps) {
+  const isInNavigationMode = useIsInNavigationMode()
   const selectedBaseLayer = useMainAppSelector(state => state.map.selectedBaseLayer)
 
   const tileCacheMapRef = useRef(new Map<string, string>())
@@ -56,12 +58,15 @@ function UnmemoizedBaseLayer({ map }: BaseLayerProps) {
           }),
           zIndex: 0
         }),
+      /**
+       * Only the LIGHT layer is cached using the service worker, in navigation mode
+       */
       LIGHT: () =>
         new TileLayer({
           className: LayerProperties.BASE_LAYER.code,
           source: new XYZ({
-            maxZoom: 19,
-            tileLoadFunction: loadTileFromCacheOrFetch,
+            maxZoom: isInNavigationMode ? 11 : 19,
+            tileLoadFunction: isInNavigationMode ? undefined : loadTileFromCacheOrFetch,
             urls: ['a', 'b', 'c', 'd'].map(
               subdomain => `https://${subdomain}.basemaps.cartocdn.com/light_all/{z}/{x}/{y}.png`
             )
@@ -100,7 +105,7 @@ function UnmemoizedBaseLayer({ map }: BaseLayerProps) {
           zIndex: 0
         })
     }),
-    [loadTileFromCacheOrFetch]
+    [loadTileFromCacheOrFetch, isInNavigationMode]
   )
 
   useEffect(() => {
