+++ frontend/src/workers/settings.ts
@@ -1,1 +1,4 @@
+/**
+ * /!\ The service worker filename is specified in the package.jon's `bundle-sw` command
+ */
 export const SERVICE_WORKER_PATH = `${process.env.PUBLIC_URL}/service-worker.js`
